# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "MJ Ma",
            "url": "https://www.linkedin.com/in/mj-profile/",
            "email": "m.mjw.ma@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/jwma/jump-jump/blob/master/LICENSE"
        }
    },
    "paths": {
...
        "/short-link/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "短链接列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "短链接"
                ],
                "summary": "短链接列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.ListShortLinksAPIResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建短链接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "短链接"
                ],
                "summary": "创建短链接",
                "parameters": [
                    {
                        "description": "创建短链接请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateShortLinkAPIRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.CreateShortLinkAPIResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/short-link/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取指定 ID 短链接详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "短链接"
                ],
                "summary": "获取指定 ID 短链接",
                "parameters": [
                    {
                        "type": "string",
                        "description": "短链接 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.GetShortLinkAPIResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除短链接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "短链接"
                ],
                "summary": "删除短链接",
                "parameters": [
                    {
                        "type": "string",
                        "description": "短链接 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "更新短链接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "短链接"
                ],
                "summary": "更新短链接",
                "parameters": [
                    {
                        "type": "string",
                        "description": "短链接 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新短链接请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateShortLinkAPIRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.UpdateShortLinkAPIResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/short-link/{id}/data": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "可查询短链接某个日期范围内的访问数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "短链接"
                ],
                "summary": "短链接访问数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "短链接 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "开始日期 YYYY-mm-dd",
                        "name": "startDate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "结束日期 YYYY-mm-dd",
                        "name": "endDate",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.ShortLinkDataAPIResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateShortLinkAPIRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "idLength": {
                    "type": "integer"
                },
                "isEnable": {
                    "type": "boolean"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "models.CreateShortLinkAPIResponseData": {
            "type": "object",
            "properties": {
                "shortLink": {
                    "$ref": "#/definitions/models.ShortLinkData"
                }
            }
        },
...
        "models.DailyStats": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "pv": {
                    "type": "integer"
                },
                "uv": {
                    "type": "integer"
                }
            }
        },
...
        "models.GetShortLinkAPIResponseData": {
            "type": "object",
            "properties": {
                "shortLink": {
                    "$ref": "#/definitions/models.ShortLinkData"
                }
            }
        },
...
        "models.ListShortLinksAPIResponseData": {
            "type": "object",
            "properties": {
                "shortLinks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ShortLinkData"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "models.RequestHistory": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "os": {
                    "type": "string"
                },
                "shortLinkId": {
                    "type": "string"
                },
                "tenantId": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "ua": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "models.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int",
                    "default": 0,
                    "example": 0
                },
                "data": {},
                "msg": {
                    "type": "string",
                    "default": "ok",
                    "example": "ok"
                }
            }
        },
        "models.ShortLinkData": {
            "type": "object",
            "properties": {
                "createTime": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isEnable": {
                    "type": "boolean"
                },
                "updateTime": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "models.ShortLinkDataAPIResponseData": {
            "type": "object",
            "properties": {
                "daily": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.DailyStats"
                    }
                },
                "histories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.RequestHistory"
                    }
                },
                "osDist": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "models.UpdateShortLinkAPIRequest": {
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "isEnable": {
                    "type": "boolean"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "models.UpdateShortLinkAPIResponseData": {
            "type": "object",
            "properties": {
                "shortLink": {
                    "$ref": "#/definitions/models.ShortLinkData"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}